requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
dependency-injector==4.43.0
fastapi==0.115.5
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
magic-filter==1.0.12
Mako==1.3.7
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
yarl==1.18.0
